daily-reports/requirements.txt
# Required Python packages for Markdown to LaTeX to PDF conversion
# Install using: pip install -r requirements.txt

# Core dependencies
python-dotenv>=0.19.0  # For environment variable management if needed in the future
